#334: Move other helpers to helpers folder
