+++ CourageScores/ClientApp/src/components/division_fixtures/DivisionFixtureDate.js
@@ -1,4 +1,4 @@
-import {getFilters} from "./FilterUtilities";
+import {getFilters} from "../../helpers/filters";
 import {any, isEmpty} from "../../helpers/collections";
 import {renderDate} from "../../helpers/rendering";
 import {FixtureDateNote} from "./FixtureDateNote";
