+++ CourageScores/ClientApp/src/components/division_fixtures/DivisionFixtures.js
@@ -7,7 +7,7 @@ import {stateChanged} from "../../helpers/events";
 import {useApp} from "../../AppContainer";
 import {useDivisionData} from "../DivisionDataContainer";
 import {DivisionFixtureDate} from "./DivisionFixtureDate";
-import {changeFilter, initFilter} from "./FilterUtilities";
+import {changeFilter, initFilter} from "../../helpers/filters";
 import {Dialog} from "../common/Dialog";
 
 export function DivisionFixtures({ setNewFixtures }) {
