+++ CourageScores/ClientApp/src/components/division_fixtures/scores/Score.js
@@ -24,7 +24,7 @@ import {Dialog} from "../../common/Dialog";
 import {EditPlayerDetails} from "../../division_players/EditPlayerDetails";
 import {LeagueFixtureContainer} from "../LeagueFixtureContainer";
 import {MatchTypeContainer} from "./MatchTypeContainer";
-import {getMatchDefaults, getMatchOptionDefaults, getMatchOptionsLookup} from "./MatchOptionHelpers";
+import {getMatchDefaults, getMatchOptionDefaults, getMatchOptionsLookup} from "../../../helpers/matchOptions";
 import {PageError} from "../../PageError";
 
 export function Score() {
