+++ CourageScores/ClientApp/src/helpers/build.test.js
@@ -2,6 +2,8 @@
 
 describe('build', () => {
     describe('getBuildDetail', () => {
-        // tested via App.test.js as it requires access to the DOM
+        it('does nothing', async () => {
+            // tested via App.test.js as it requires access to the DOM
+        });
     });
 });
