+++ CourageScores/ClientApp/src/helpers/collections.test.js
@@ -1,3 +1,5 @@
+// noinspection JSUnresolvedReference
+
 import {
     all,
     any,
