+++ CourageScores/ClientApp/src/helpers/tournaments.js
